Remove parallel builds
Parallel builds are really challenging on OSX for some reason. See #7.
Since it doesn't appear like this is going to be addressed in a user
friendly way anytime soon, comment the parallel build line out. It only
happens when compiling deps for the first time, so the build time is
annoying, but not an every build thing.

Fixes #7.